+++ copy.go
@@ -17,7 +17,6 @@ var (
 	errBinaryCopyNotSupported     = errors.New("pq: only text format supported for COPY")
 	errCopyToNotSupported         = errors.New("pq: COPY TO is not supported")
 	errCopyNotSupportedOutsideTxn = errors.New("pq: COPY is only allowed inside a transaction")
-	errCopyInProgress             = errors.New("pq: COPY in progress")
 )
 
 // CopyIn creates a COPY FROM statement which can be prepared with Tx.Prepare().
@@ -368,7 +367,7 @@ func (ci *copyin) Close() error {
 	}
 
 	<-ci.done
-	ci.cn.inCopy = false
+	ci.cn.inProgress.Store(false)
 
 	if err := ci.err(); err != nil {
 		return err